Crate rendy

Rendy's top level crate. Reexports all others.

Re-exports

pub use crate::core::hal;

Modules

command
This crate revolves around command recording and submission.

core
Crate that contains utility modules used by other rendy crates

descriptor

factory
Higher-level rendy interface.

frame
Everything that is frame related.

graph
Framegraph implementation for Rendy engine.

init
Simple rendy initialization. Takes most bolierplate required for init rendy on different platforms/backends. It is still possible to construct everything manually if your case is not supported by this module.

memory
GPU memory management

mesh
This crates provides means to deal with vertex buffers and meshes.

resource
This crate provide methods to create/destroy and otherwise manage device resources. Primarily focus on buffers and images.

shader
Shader compilation.

texture
Texture creation and usage

wsi
Window system integration.

Macros

with_any_rendy
Init rendy and execute code based on chosen backend

with_any_windowed_rendy
Init rendy and execute code based on chosen backend