Struct rendy::command::CommandBuffer

pub struct CommandBuffer<B, C, S, L = PrimaryLevel, R = NoIndividualReset> where
    B: Backend,  { /* fields omitted */ }

Command buffer wrapper. This wrapper defines state with usage, level and ability to be individually reset at type level. This way many methods become safe.

Implementations

impl<B, C, U, L, R> CommandBuffer<B, C, RecordingState<U, OutsideRenderPass>, L, R> where
B: Backend,
C: Capability,
L: Level,

pub fn encoder(&mut self) -> Encoder<'_, B, C, L>

Get encoder that will encode commands into this command buffer.

impl<B, C, U, R> CommandBuffer<B, C, RecordingState<U, RenderPassContinue>, SecondaryLevel, R> where
B: Backend,
C: Supports<Graphics>,

pub fn render_pass_encoder(&mut self) -> RenderPassEncoder<'_, B>

Get encoder that will encode render-pass commands into this command buffer.

impl<B, C, P, L, R> CommandBuffer<B, C, ExecutableState<OneShot, P>, L, R> where
B: Backend,
L: Copy,
P: Copy,

pub fn submit_once(
self
) -> (Submit<B, NoSimultaneousUse, L, P>, CommandBuffer<B, C, PendingState<InvalidState>, L, R>)

Produce Submit object that can be used to populate submission.

impl<B, C, S, L, P, R> CommandBuffer<B, C, ExecutableState<MultiShot<S>, P>, L, R> where
B: Backend,
L: Copy,
P: Copy,
S: Copy,

pub fn submit(
self
) -> (Submit<B, S, L, P>, CommandBuffer<B, C, PendingState<ExecutableState<MultiShot<S>, P>>, L, R>)

Produce Submit object that can be used to populate submission.

impl<B, C, S, L, R> CommandBuffer<B, C, S, L, R> where
B: Backend,

pub unsafe fn change_state<U>(
self,
f: impl FnOnce(S) -> U
) -> CommandBuffer<B, C, U, L, R>

Change state of the command buffer.

Safety

This method must be used only to reflect state changed due to raw handle usage.

pub fn capability(&self) -> C where
C: Capability,

Get buffers capability.

pub fn family(&self) -> FamilyId

Get buffers family.

pub fn with_queue_type(self) -> CommandBuffer<B, QueueType, S, L, R> where
C: Capability,

Convert capability level.

pub fn with_capability<U>(
self
) -> Result<CommandBuffer<B, U, S, L, R>, CommandBuffer<B, C, S, L, R>> where
C: Supports<U>,

Convert capability level.

impl<B, C, L, R> CommandBuffer<B, C, InitialState, L, R> where
B: Backend,

pub fn begin<'a, U, P>(
self,
usage: U,
info: impl BeginInfo<'a, B, L, PassRelation = P>
) -> CommandBuffer<B, C, RecordingState<U, P>, L, R> where
P: RenderPassRelation<L>,
U: Usage,

Begin recording command buffer.

Parameters

usage - specifies usage of the command buffer. Possible types are OneShot, MultiShot.

impl<'a, B, C, U, P, L, R> CommandBuffer<B, C, RecordingState<U, P>, L, R> where
B: Backend,

pub fn finish(self) -> CommandBuffer<B, C, ExecutableState<U, P>, L, R>

Finish recording command buffer.

impl<B, C, N, L, R> CommandBuffer<B, C, PendingState<N>, L, R> where
B: Backend,

pub unsafe fn mark_complete(self) -> CommandBuffer<B, C, N, L, R>

Mark command buffer as complete.

Safety

None of Submit instances created from this CommandBuffer are alive.

If this is PrimaryLevel buffer then for each command queue where Submit instance (created from this CommandBuffer) was submitted at least one Fence submitted within same Submission or later in unset state was set.

If this is Secondary buffer then all primary command buffers where Submit instance (created from this CommandBuffer) was submitted must be complete.

impl<B, C, S, L> CommandBuffer<B, C, S, L, IndividualReset> where
B: Backend,
S: Resettable,

pub fn reset(self) -> CommandBuffer<B, C, InitialState, L, IndividualReset>

Reset command buffer.

impl<B, C, S, L> CommandBuffer<B, C, S, L, NoIndividualReset> where
B: Backend,
S: Resettable,

pub unsafe fn mark_reset(
self
) -> CommandBuffer<B, C, InitialState, L, NoIndividualReset>

Mark command buffer as reset.

Safety

This function must be used only to reflect command buffer being reset implicitly. For instance:

CommandPool::reset on pool from which the command buffer was allocated.

Raw handle usage.

impl<B, C, S, L, R> CommandBuffer<B, C, S, L, R> where
B: Backend,
S: Resettable,

pub fn into_raw(self) -> <B as Backend>::CommandBuffer

Dispose of command buffer wrapper releasing raw comman buffer value. This function is intended to be used to deallocate command buffer.

pub fn raw(&mut self) -> &mut <B as Backend>::CommandBuffer

Get raw command buffer handle.

impl<B, C, S, L, R> CommandBuffer<B, C, S, L, R> where
B: Backend,

pub fn family_id(&self) -> FamilyId

Get owner id.

pub fn assert_family_owner(&self, family: &Family<B, C>)

Assert specified family is owner.

pub fn assert_device_owner(&self, device: &Device<B>)

Assert specified device is owner.

pub fn assert_instance_owner(&self, instance: &Instance<B>)

Assert specified instance is owner.