Struct rendy::command::Encoder

pub struct Encoder<'a, B, C, L> where
    B: Backend,  { /* fields omitted */ }

Trait to encode commands outside render pass.

Implementations

impl<'a, B, C> Encoder<'a, B, C, PrimaryLevel> where
B: Backend,

pub fn begin_render_pass_inline(
&mut self,
render_pass: &<B as Backend>::RenderPass,
framebuffer: &<B as Backend>::Framebuffer,
render_area: Rect,
clear_values: &[ClearValueRaw]
) -> RenderPassInlineEncoder<B> where
C: Supports<Graphics>,

Beging recording render pass inline.

pub fn begin_render_pass_secondary(
&mut self,
render_pass: &<B as Backend>::RenderPass,
framebuffer: &<B as Backend>::Framebuffer,
render_area: Rect,
clear_values: &[ClearValueRaw]
) -> RenderPassSecondaryEncoder<B> where
C: Supports<Graphics>,

Beging recording render pass secondary.

pub fn execute_commands(
&mut self,
submittables: impl IntoIterator<Item = impl Submittable<B, SecondaryLevel, OutsideRenderPass>>
)

Execute commands from secondary buffers.

impl<'a, B, C, L> Encoder<'a, B, C, L> where
B: Backend,

pub fn level(&self) -> L where
L: Level,

Get encoder level.

pub unsafe fn copy_buffer(
&mut self,
src: &<B as Backend>::Buffer,
dst: &<B as Backend>::Buffer,
regions: impl IntoIterator<Item = BufferCopy>
) where
C: Supports<Transfer>,

Copy buffer regions. src and dst can be the same buffer or alias in memory. But regions must not overlap. Otherwise resulting values are undefined.

Safety

The size of the copy region in any regions must not exceed the length of the corresponding buffer.

See: https://www.khronos.org/registry/vulkan/specs/1.1-extensions/man/html/vkCmdCopyBuffer.html

pub unsafe fn copy_buffer_to_image(
&mut self,
src: &<B as Backend>::Buffer,
dst: &<B as Backend>::Image,
dst_layout: Layout,
regions: impl IntoIterator<Item = BufferImageCopy>
) where
C: Supports<Transfer>,

Copy buffer region to image subresource range.

Safety

Same as copy_buffer()

See: https://www.khronos.org/registry/vulkan/specs/1.1-extensions/man/html/vkCmdCopyBufferToImage.html

pub unsafe fn copy_image(
&mut self,
src: &<B as Backend>::Image,
src_layout: Layout,
dst: &<B as Backend>::Image,
dst_layout: Layout,
regions: impl IntoIterator<Item = ImageCopy>
) where
C: Supports<Transfer>,

Copy image regions.

Safety

Same as copy_buffer()

See: https://www.khronos.org/registry/vulkan/specs/1.1-extensions/man/html/vkCmdCopyImage.html

pub unsafe fn blit_image(
&mut self,
src: &<B as Backend>::Image,
src_layout: Layout,
dst: &<B as Backend>::Image,
dst_layout: Layout,
filter: Filter,
regions: impl IntoIterator<Item = ImageBlit>
) where
C: Supports<Graphics>,

Blit image regions, potentially using specified filter when resize is necessary.

Safety

Same as copy_buffer()

See: https://www.khronos.org/registry/vulkan/specs/1.1-extensions/man/html/vkCmdBlitImage.html

pub unsafe fn dispatch(&mut self, x: u32, y: u32, z: u32) where
C: Supports<Compute>,

Dispatch compute.

Safety

See: https://www.khronos.org/registry/vulkan/specs/1.1-extensions/man/html/vkCmdDispatch.html

pub unsafe fn dispatch_indirect(
&mut self,
buffer: &<B as Backend>::Buffer,
offset: u64
) where
C: Supports<Compute>,

Dispatch indirect. Similar to dispatch except takes vertices and indices from buffer at specified offset. buffer must contain DispatchCommand at offset.

Safety

See: https://www.khronos.org/registry/vulkan/specs/1.1-extensions/man/html/vkCmdDispatchIndirect.html