Struct rendy::command::Family

pub struct Family<B, C = QueueType> where
    B: Backend,  { /* fields omitted */ }

Family of the command queues. Queues from one family can share resources and execute command buffers associated with the family. All queues of the family have same capabilities.

Methods

impl<B> Family<B, QueueType> where
B: Backend,

pub unsafe fn from_device<impl gfx_hal::queue::QueueFamily>(
queues: &mut Queues<B>,
index: QueueFamilyId,
count: usize,
family: &impl gfx_hal::queue::QueueFamily
) -> Family<B, QueueType> where
impl gfx_hal::queue::QueueFamily: QueueFamily,

Query queue family from device.

Safety

This function shouldn't be used more then once with the same parameters. Raw queue handle queried from device can make Family usage invalid. family must be one of the family indices used during device creation. properties must be the properties retuned for queue family from physical device.

impl<B, C> Family<B, C> where
B: Backend,

pub fn index(&self) -> QueueFamilyId

Get id of the family.

pub fn queues(
&self
) -> &[impl RawCommandQueue<B>]

Get queues of the family.

pub fn queues_mut(
&mut self
) -> &mut [impl RawCommandQueue<B>]

Get queues of the family.

pub unsafe fn submit<'a, impl std::borrow::Borrow + 'a, impl IntoIterator + 'a), gfx_hal::pso::PipelineStage)>, impl Submittable, impl IntoIterator>, impl std::borrow::Borrow + 'a, impl IntoIterator + 'a)>, impl IntoIterator + 'a), gfx_hal::pso::PipelineStage)>, impl IntoIterator>, impl IntoIterator + 'a)>>>>(
&mut self,
queue: usize,
submissions: impl IntoIterator + 'a), gfx_hal::pso::PipelineStage)>, impl IntoIterator>, impl IntoIterator + 'a)>>>,
fence: Option<&<B as Backend>::Fence>
) where
impl IntoIterator + 'a), gfx_hal::pso::PipelineStage)>: IntoIterator<Item = (&'a impl std::borrow::Borrow + 'a, PipelineStage)>,
impl IntoIterator + 'a), gfx_hal::pso::PipelineStage)>, impl IntoIterator>, impl IntoIterator + 'a)>>>: IntoIterator<Item = Submission<B, impl IntoIterator + 'a), gfx_hal::pso::PipelineStage)>, impl IntoIterator>, impl IntoIterator + 'a)>>>,
impl IntoIterator + 'a)>: IntoIterator<Item = &'a impl std::borrow::Borrow + 'a>,
impl IntoIterator>: IntoIterator<Item = impl Submittable>,
impl Submittable: Submittable<B, PrimaryLevel, OutsideRenderPass>,
impl std::borrow::Borrow + 'a: 'a + Borrow<<B as Backend>::Semaphore> + 'a + Borrow<<B as Backend>::Semaphore>,

Submit commands to the queue of the family.

pub fn create_pool<R, impl gfx_hal::Device>(
&self,
device: &impl gfx_hal::Device
) -> Result<CommandPool<B, C, R>, OutOfMemory> where
C: Capability,
R: Reset,
impl gfx_hal::Device: Device<B>,

Create command pool associated with the family. Command buffers created from the pool could be submitted to the queues of the family.

pub fn capability(&self) -> C where
C: Capability,

Get family capability.

pub fn dispose(self)

Dispose of queue family container.

pub fn with_queue_type(self) -> Family<B, QueueType> where
C: Capability,

Convert capability from type-level to value-level.

pub fn with_capability<U>(self) -> Result<Family<B, U>, Family<B, C>> where
C: Supports<U>,

Convert capability into type-level one.