Module rendy::core::hal::prelude

Prelude module re-exports all the traits necessary to use gfx-hal.

Traits

_
Represents a physical device (such as a GPU) capable of supporting the given backend.

_
A trait that describes all the operations that must be provided by a Backend's command buffer.

_
Overview

_
The allocated command buffers are associated with the creating command queue.

_
RawCommandQueue are abstractions to the internal GPU execution engines. Commands are executed on the the device by submitting command buffers to queues.

_
General information about a queue family, available upon adapter discovery.

_
A surface trait that exposes the ability to present images on the associtated swap chain.

_
A Surface abstracts the surface of a native window.

_
The Swapchain is the backend representation of the surface. It consists of multiple buffers, which will be presented on the surface.

_
An instantiated backend.

_
A descriptor pool is a collection of memory from which descriptor sets are allocated.