Trait rendy::core::types::vertex::AsAttributes

pub trait AsAttributes where
    <Self::Iter as Iterator>::Item == Attribute, {
    type Iter: Iterator;
    fn attributes(self) -> Self::Iter;
}

Represent types that can be interpreted as list of vertex attributes.

Associated Types

type Iter: Iterator

The iterator type for retreived attributes

Loading content...
Required methods

fn attributes(self) -> Self::Iter

Retreive a list of vertex attributes with offsets relative to beginning of that list

Loading content...