Module rendy::factory

Higher-level rendy interface.

Structs

Barriers
Handles combining multiple image and buffer barriers that must be made before and after some target operations.

BasicDevicesConfigure
Basics adapters config.

BasicHeapsConfigure
Basic heaps config. It uses some arbitrary values that can be considered sane default for today (year 2019) hardware and software.

BlitImageState
A region and image states for one image in a blit.

BlitRegion
A region to be blitted including the source and destination images and states,

Blitter
Manages blitting images across families and queues.

BufferState
State of the buffer on device.

Config
Factory initialization config.

Factory
Higher level device interface. Manges memory, resources and queue families.

ImageState
State of the image on device.

OneGraphicsQueue
QueuePicker that picks first graphics queue family.

SavedHeapsConfig
Saved config for allocators. This config can be loaded from config files in any format supported by serde ecosystem.

SavedQueueConfig
Saved config for queues. This config can be loaded from config files in any format supported by serde ecosystem.

Enums

ImageStateOrLayout
Either image state or just layout for image that is not used by device.

UploadError
Failure uploading a buffer or an image.

Traits

DevicesConfigure
Devices configuration. Picks physical device to use.

HeapsConfigure
Heaps configuration.

QueuesConfigure
Queues configuration.

Functions

blit_image⚠
Blits one or more regions from src_image into dst_image using specified Filter

init_with_instance
Initialize Factory and Queue Families associated with Device using existing Instance.

init_with_instance_ref
Initialize Factory and Queue Families associated with Device using existing Instance.