Struct rendy::factory::Barriers

pub struct Barriers<B> where
    B: Backend,  { /* fields omitted */ }

Handles combining multiple image and buffer barriers that must be made before and after some target operations.

Implementations

impl<B> Barriers<B> where
B: Backend,

pub fn new(
target_stages: PipelineStage,
target_buffer_access: Access,
target_image_access: Access
) -> Barriers<B>

Create a new Barriers instance with target stages and accesses

pub fn add_image(
&mut self,
image: Handle<Image<B>>,
image_range: SubresourceRange,
last_stage: PipelineStage,
last_access: Access,
last_layout: Layout,
target_layout: Layout,
next_stage: PipelineStage,
next_access: Access,
next_layout: Layout
)

Add an image to the barriers

pub fn add_buffer(
&mut self,
last_stage: PipelineStage,
last_access: Access,
next_stage: PipelineStage,
next_access: Access
)

Add a buffer to the barriers

pub fn encode_before<C, L>(&mut self, encoder: &mut Encoder<'_, B, C, L>)

Encode the barriers that should come before the target operations

pub fn encode_after<C, L>(&mut self, encoder: &mut Encoder<'_, B, C, L>)

Encode the barriers that should come after the target operations