Struct rendy::factory::Factory

pub struct Factory<B> where
    B: Backend,  { /* fields omitted */ }

Higher level device interface. Manges memory, resources and queue families.

Methods

impl<B> Factory<B> where
B: Backend,

pub fn init<impl gfx_hal::Instance, impl HeapsConfigure, impl QueuesConfigure>(
instance: impl gfx_hal::Instance,
config: Config<impl HeapsConfigure, impl QueuesConfigure>
) -> Result<Factory<B>, Error> where
impl HeapsConfigure: HeapsConfigure,
impl QueuesConfigure: QueuesConfigure,
impl gfx_hal::Instance: Instance<Backend = B>,

Creates a new Factory based off of a Config<Q, W> with some QueuesConfigure from the specified vk::PhysicalDevice.

pub fn wait_idle(&self) -> Result<(), HostExecutionError>

Wait for whole device become idle. This function is very heavy and usually used only for teardown.

pub fn dispose(self)

Dispose of the Factory.

pub fn create_buffer<impl buffer::Usage>(
&self,
align: u64,
size: u64,
usage: impl buffer::Usage
) -> Result<Buffer<B>, Error> where
impl buffer::Usage: Usage,

Creates a buffer that is managed with the specified properties.

pub fn create_image<impl image::Usage>(
&self,
align: u64,
kind: Kind,
levels: u8,
format: Format,
tiling: Tiling,
view_caps: ViewCapabilities,
usage: impl image::Usage
) -> Result<Image<B>, Error> where
impl image::Usage: Usage,

Creates an image that is mananged with the specified properties.

pub unsafe fn upload_buffer<T>(
&self,
buffer: &mut Buffer<B>,
offset: u64,
content: &[T],
family: QueueFamilyId,
_access: Access
) -> Result<(), Error> where
T: Copy,

Upload buffer content.

Safety

Buffer must be created by this Factory.

Caller must ensure that device won't write to or read from the memory region occupied by this buffer.

pub unsafe fn upload_visible_buffer<T>(
&self,
buffer: &mut Buffer<B>,
offset: u64,
content: &[T]
) -> Result<(), Error> where
T: Copy,

Update buffer bound to host visible memory.vk::AccessFlags.

Safety

Caller must ensure that device won't write to or read from the memory region occupied by this buffer.

pub fn upload_image<T>(
&self,
_image: &mut Image<B>,
_data_with: u32,
_data_height: u32,
_image_layers: SubresourceLayers,
_image_offset: Offset,
_image_extent: Extent,
_content: &[T],
_family: QueueFamilyId,
_access: Access,
_layout: Layout
) -> Result<(), Error> where
T: Copy,

Upload image.

pub fn create_surface(&self, window: Window) -> Surface<B>

Create rendering surface from window.

pub fn create_target(
&self,
surface: Surface<B>,
image_count: u32,
usage: Usage
) -> Result<Target<B>, Error>

Create target out of rendering surface.

pub unsafe fn destroy_target(&self, target: Target<B>) -> Window

Destroy target returning underlying window back to the caller.

pub fn families(&self) -> &[Family<B, QueueType>]

Get queue families of the factory.

pub fn families_mut(&mut self) -> &mut [Family<B, QueueType>]

Get queue families of the factory.

pub fn family(&self, index: QueueFamilyId) -> &Family<B, QueueType>

Get queue families of the factory.

pub fn family_mut(&mut self, index: QueueFamilyId) -> &mut Family<B, QueueType>

Get queue families of the factory.

pub fn target_support(&self, family: QueueFamilyId, target: &Target<B>) -> bool

Get surface support for family.

pub fn device(&self) -> &impl Device<B>

Get device.

pub fn physical(&self) -> &impl PhysicalDevice<B>

Get physical device.

pub fn create_semaphore(&self) -> Result<<B as Backend>::Semaphore, OutOfMemory>

Create new semaphore

pub fn destroy_semaphore(&self, semaphore: <B as Backend>::Semaphore)

Destroy semaphore

pub fn create_fence(
&self,
signaled: bool
) -> Result<<B as Backend>::Fence, OutOfMemory>

Create new fence

pub fn reset_fence(
&self,
fence: &<B as Backend>::Fence
) -> Result<(), OutOfMemory>

Wait for the fence become signeled.

pub fn reset_fences<impl std::borrow::Borrow, impl IntoIterator>>(
&self,
fences: impl IntoIterator>
) -> Result<(), OutOfMemory> where
impl IntoIterator>: IntoIterator<Item = impl std::borrow::Borrow>,
impl std::borrow::Borrow: Borrow<<B as Backend>::Fence>,

Wait for the fence become signeled.

pub fn wait_for_fence(
&self,
fence: &<B as Backend>::Fence,
timeout_ns: u64
) -> Result<bool, OomOrDeviceLost>

Wait for the fence become signeled.

pub fn wait_for_fences<impl std::borrow::Borrow, impl IntoIterator>>(
&self,
fences: impl IntoIterator>,
wait_for: WaitFor,
timeout_ns: u64
) -> Result<bool, OomOrDeviceLost> where
impl IntoIterator>: IntoIterator<Item = impl std::borrow::Borrow>,
impl std::borrow::Borrow: Borrow<<B as Backend>::Fence>,

Wait for the fences become signeled.

pub unsafe fn destroy_fence(&self, fence: <B as Backend>::Fence)

Destroy fence.

pub fn create_command_pool<R>(
&self,
family: QueueFamilyId
) -> Result<CommandPool<B, QueueType, R>, Error> where
R: Reset,

Create new command pool for specified family.

pub unsafe fn destroy_command_pool<R>(&self, pool: CommandPool<B, QueueType, R>) where
R: Reset,

Create new command pool for specified family.

impl<B> Factory<B> where
B: Backend,

pub fn new<impl HeapsConfigure, impl QueuesConfigure>(
config: Config<impl HeapsConfigure, impl QueuesConfigure>
) -> Result<Factory<B>, Error> where
impl HeapsConfigure: HeapsConfigure,
impl QueuesConfigure: QueuesConfigure,

Init factory.