Trait rendy::factory::HeapsConfigure

pub unsafe trait HeapsConfigure where
    <Self::Types as IntoIterator>::Item == (Properties, u32, HeapsConfig),
    <Self::Heaps as IntoIterator>::Item == u64, {
    type Types: IntoIterator;
    type Heaps: IntoIterator;
    fn configure(
        &self, 
        properties: &MemoryProperties
    ) -> (Self::Types, Self::Heaps);
}

Heaps configuration.

Method configure receives memory properties and emits iterator memory types together with configurations for allocators and iterator over heaps sizes.

Associated Types

type Types: IntoIterator

Iterator over memory types.

type Heaps: IntoIterator

Iterator over heaps.

Loading content...
Required methods

fn configure(&self, properties: &MemoryProperties) -> (Self::Types, Self::Heaps)

Configure.

Loading content...