Module rendy::graph

Framegraph implementation for Rendy engine.

Modules

present
Defines present node.

render
Defines render pass node.

Structs

BufferAccess
Buffer access node will perform. Node must not perform any access to the buffer not specified in access. All access must be between logically first and last stages.

BufferBarrier
Buffer pipeline barrier.

BufferId
Id of the buffer in graph.

Graph
Graph that renders whole frame.

GraphBuilder
Build graph from nodes and resource.

ImageAccess
Image access node wants to perform.

ImageBarrier
Image pipeline barrier. Node implementation must insert it before first command that uses the image. Barrier must be inserted even if this node doesn't use the image.

ImageId
Id of the image (or target) in graph.

NodeBuffer
Buffer shared between nodes.

NodeBuilder
Builder for the node.

NodeId
Id of the node in graph.

NodeImage
Image shared between nodes.

Traits

AnyNode
Trait-object safe Node.

AnyNodeDesc
Trait-object safe NodeDesc.

Node
The node is building block of the framegraph. Node defines set of resources and operations to perform over them. Read-only data for operations comes from auxiliary data source T.

NodeDesc
Builder of the node. Implementation of the builder type provide framegraph with static information about node that is used for building the node.

NodeSubmittable
NodeSubmittable

Functions

gfx_acquire_barriers
Convert graph barriers into gfx barriers.

gfx_release_barriers
Convert graph barriers into gfx barriers.