Struct rendy::graph::DescBuilder

pub struct DescBuilder<B, T, N> where
    B: Backend,
    T: ?Sized,  { /* fields omitted */ }

Builder for the node.

Methods

impl<B, T, N> DescBuilder<B, T, N> where
B: Backend,
T: ?Sized,

pub fn new(desc: N) -> DescBuilder<B, T, N>

Create new builder out of desc

pub fn add_buffer(&mut self, buffer: BufferId) -> &mut DescBuilder<B, T, N>

Add buffer to the node. This method must be called for each buffer node uses.

pub fn with_buffer(self, buffer: BufferId) -> DescBuilder<B, T, N>

Add buffer to the node. This method must be called for each buffer node uses.

pub fn add_image(&mut self, image: ImageId) -> &mut DescBuilder<B, T, N>

Add image to the node. This method must be called for each image node uses.

pub fn with_image(self, image: ImageId) -> DescBuilder<B, T, N>

Add image to the node. This method must be called for each image node uses.

pub fn add_dependency(
&mut self,
dependency: NodeId
) -> &mut DescBuilder<B, T, N>

Add dependency. Node will be placed after its dependencies.

pub fn with_dependency(self, dependency: NodeId) -> DescBuilder<B, T, N>

Add dependency. Node will be placed after its dependencies.