Trait rendy::graph::Node

pub trait Node<B, T>: 'static + for<'a> NodeSubmittable<'a, B> + Send + Sync + Debug where
    B: Backend,
    T: ?Sized, {
    type Capability: Capability;
    fn run(
        &'a mut self, 
        ctx: &GraphContext<B>, 
        factory: &Factory<B>, 
        aux: &T, 
        frames: &'a Frames<B>
    ) -> Self::Submittables;

    unsafe fn dispose(self, factory: &mut Factory<B>, aux: &T);
}

The node is building block of the framegraph. Node defines set of resources and operations to perform over them. Read-only data for operations comes from auxiliary data source T.

Parameters

B - backend type. T - auxiliary data type.

Associated Types

type Capability: Capability

Capability required by node. Graph will execute this node on command queue that supports this capability level.

Loading content...
Required methods

fn run(
&'a mut self,
ctx: &GraphContext<B>,
factory: &Factory<B>,
aux: &T,
frames: &'a Frames<B>
) -> Self::Submittables

Record commands required by node. Returned submits are guaranteed to be submitted within specified frame.

unsafe fn dispose(self, factory: &mut Factory<B>, aux: &T)

Dispose of the node.

Safety

Must be called after waiting for device idle.

Loading content...