Trait rendy::mesh::WithAttribute

pub trait WithAttribute<F>: AsVertex where
    F: AsAttribute, {
    const ATTRIBUTE: Element<Format>;
}

Trait implemented by all valid vertex formats for each field

Associated Constants

const ATTRIBUTE: Element<Format>

Individual format of the attribute for this vertex format