Struct rendy::resource::Resources

pub struct Resources<B> where
    B: Backend,  { /* fields omitted */ }

Resource manager. It can be used to create and destroy resources such as buffers and images.

Methods

impl<B> Resources<B> where
B: Backend,

pub fn new() -> Resources<B>

Create new Resources instance.

pub fn create_buffer<impl gfx_hal::Device, impl buffer::Usage>(
&self,
device: &impl gfx_hal::Device,
heaps: &mut Heaps<B>,
align: u64,
size: u64,
usage: impl buffer::Usage
) -> Result<Buffer<B>, Error> where
impl buffer::Usage: Usage,
impl gfx_hal::Device: Device<B>,

Create a buffer and bind to the memory that support intended usage.

pub fn destroy_buffer(&mut self, buffer: Buffer<B>)

Destroy buffer. Buffer can be dropped but this method reduces overhead.

pub fn create_image<impl gfx_hal::Device, impl image::Usage>(
&self,
device: &impl gfx_hal::Device,
heaps: &mut Heaps<B>,
align: u64,
kind: Kind,
levels: u8,
format: Format,
tiling: Tiling,
view_caps: ViewCapabilities,
usage: impl image::Usage
) -> Result<Image<B>, Error> where
impl gfx_hal::Device: Device<B>,
impl image::Usage: Usage,

Create an image and bind to the memory that support intended usage.

pub fn destroy_image(&mut self, image: Image<B>)

Destroy image. Image can be dropped but this method reduces overhead.

pub unsafe fn cleanup<impl gfx_hal::Device>(
&mut self,
device: &impl gfx_hal::Device,
heaps: &mut Heaps<B>
) where
impl gfx_hal::Device: Device<B>,

Recycle dropped resources.

Safety

Device must not attempt to use previously dropped buffers and images.