Module rendy::texture

Texture creation and usage

Modules

pixel
Typed pixel formats. More information on these can be found here

Structs

Texture
Static image. Can be loaded from various of formats.

TextureBuilder
Generics-free texture builder. Struct for staging data in preparation of building a Texture

Enums

BuildError

MipLevels
Number of mip levels

Functions

mip_levels_from_dims
Calculate the number of mip levels for a 2D image with given dimensions

Type Definitions

Rgba8Unorm
Pixel type alias.