Module rendy::texture::pixel

Typed pixel formats. More information on these can be found here

Structs

Abgr
Alpha-blue-green-red channels.

Bgr
Blue-green-red channels.

Bgra
Blue-green-red-alpha channels.

Float
Floating point representation.

Inorm
Normalized signed integer representation

Int
Signed integer representation

Iscaled
Signed scaled integer representation

Pixel
One pixel

R
Red channel.

Rg
Red-green channels.

Rgb
Red-green-blue channels.

Rgba
Red-green-blue-alpha channels.

Srgb
Unorm but in with srgb gammar correction.

Uint
Unsigned integer representation

Unorm
Normalized unsigned integer representation

Uscaled
Unsigned scaled integer representation

_8
8 bits marker type

_16
16 bits marker type

_32
32 bits marker type

_64
64 bits marker type

Traits

AsPixel
AsPixel trait for extracting the underlying data representation information from a Rust data type

ChannelRepr
Channel representation as a Rust type

ChannelSize
Byte size of each channel in the image, such as Red, Green, or other channels depending on the format.

PixelRepr
Pixel representation as a Rust type

Type Definitions

Abgr8Sint
Pixel type alias.

Abgr8Snorm
Pixel type alias.

Abgr8Srgb
Pixel type alias.

Abgr8Sscaled
Pixel type alias.

Abgr8Uint
Pixel type alias.

Abgr8Unorm
Pixel type alias.

Abgr8Uscaled
Pixel type alias.

Bgr8Sint
Pixel type alias.

Bgr8Snorm
Pixel type alias.

Bgr8Srgb
Pixel type alias.

Bgr8Sscaled
Pixel type alias.

Bgr8Uint
Pixel type alias.

Bgr8Unorm
Pixel type alias.

Bgr8Uscaled
Pixel type alias.

Bgra8Sint
Pixel type alias.

Bgra8Snorm
Pixel type alias.

Bgra8Srgb
Pixel type alias.

Bgra8Sscaled
Pixel type alias.

Bgra8Uint
Pixel type alias.

Bgra8Unorm
Pixel type alias.

Bgra8Uscaled
Pixel type alias.

R8Sint
Pixel type alias.

R8Snorm
Pixel type alias.

R8Srgb
Pixel type alias.

R8Sscaled
Pixel type alias.

R8Uint
Pixel type alias.

R8Unorm
Pixel type alias.

R8Uscaled
Pixel type alias.

R16Sint
Pixel type alias.

R16Snorm
Pixel type alias.

R16Sscaled
Pixel type alias.

R16Uint
Pixel type alias.

R16Unorm
Pixel type alias.

R16Uscaled
Pixel type alias.

R32Sfloat
Pixel type alias.

R32Sint
Pixel type alias.

R32Uint
Pixel type alias.

R64Sfloat
Pixel type alias.

R64Sint
Pixel type alias.

R64Uint
Pixel type alias.

Rg8Sint
Pixel type alias.

Rg8Snorm
Pixel type alias.

Rg8Srgb
Pixel type alias.

Rg8Sscaled
Pixel type alias.

Rg8Uint
Pixel type alias.

Rg8Unorm
Pixel type alias.

Rg8Uscaled
Pixel type alias.

Rg16Sint
Pixel type alias.

Rg16Snorm
Pixel type alias.

Rg16Sscaled
Pixel type alias.

Rg16Uint
Pixel type alias.

Rg16Unorm
Pixel type alias.

Rg16Uscaled
Pixel type alias.

Rg32Sfloat
Pixel type alias.

Rg32Sint
Pixel type alias.

Rg32Uint
Pixel type alias.

Rg64Sfloat
Pixel type alias.

Rg64Sint
Pixel type alias.

Rg64Uint
Pixel type alias.

Rgb8Sint
Pixel type alias.

Rgb8Snorm
Pixel type alias.

Rgb8Srgb
Pixel type alias.

Rgb8Sscaled
Pixel type alias.

Rgb8Uint
Pixel type alias.

Rgb8Unorm
Pixel type alias.

Rgb8Uscaled
Pixel type alias.

Rgb16Sint
Pixel type alias.

Rgb16Snorm
Pixel type alias.

Rgb16Sscaled
Pixel type alias.

Rgb16Uint
Pixel type alias.

Rgb16Unorm
Pixel type alias.

Rgb16Uscaled
Pixel type alias.

Rgb32Sfloat
Pixel type alias.

Rgb32Sint
Pixel type alias.

Rgb32Uint
Pixel type alias.

Rgb64Sfloat
Pixel type alias.

Rgb64Sint
Pixel type alias.

Rgb64Uint
Pixel type alias.

Rgba8Sint
Pixel type alias.

Rgba8Snorm
Pixel type alias.

Rgba8Srgb
Pixel type alias.

Rgba8Sscaled
Pixel type alias.

Rgba8Uint
Pixel type alias.

Rgba8Unorm
Pixel type alias.

Rgba8Uscaled
Pixel type alias.

Rgba16Sint
Pixel type alias.

Rgba16Snorm
Pixel type alias.

Rgba16Sscaled
Pixel type alias.

Rgba16Uint
Pixel type alias.

Rgba16Unorm
Pixel type alias.

Rgba16Uscaled
Pixel type alias.

Rgba32Sfloat
Pixel type alias.

Rgba32Sint
Pixel type alias.

Rgba32Uint
Pixel type alias.

Rgba64Sfloat
Pixel type alias.

Rgba64Sint
Pixel type alias.

Rgba64Uint
Pixel type alias.