Trait rendy::texture::pixel::ChannelSize

pub trait ChannelSize {
    const SIZE: u32;
}

Byte size of each channel in the image, such as Red, Green, or other channels depending on the format.

Associated Constants

const SIZE: u32

Channel representation.

Loading content...