Struct rendy::command::Encoder

pub struct Encoder<'a, B, C, L> where
    B: Backend,
    <B as Backend>::CommandBuffer: 'a,  { /* fields omitted */ }

Trait to encode commands outside render pass.

Methods

impl<'a, B, C> Encoder<'a, B, C, PrimaryLevel> where
B: Backend,

pub fn begin_render_pass_inline(
&mut self,
render_pass: &<B as Backend>::RenderPass,
framebuffer: &<B as Backend>::Framebuffer,
render_area: Rect,
clear_values: &[ClearValueRaw]
) -> RenderPassInlineEncoder<B> where
C: Supports<Graphics>,

Beging recording render pass inline.

pub fn begin_render_pass_secondary(
&mut self,
render_pass: &<B as Backend>::RenderPass,
framebuffer: &<B as Backend>::Framebuffer,
render_area: Rect,
clear_values: &[ClearValueRaw]
) -> RenderPassSecondaryEncoder<B> where
C: Supports<Graphics>,

Beging recording render pass secondary.

pub fn execute_commands<impl Submittable, impl IntoIterator>>(
&mut self,
submittables: impl IntoIterator>
) where
impl IntoIterator>: IntoIterator<Item = impl Submittable>,
impl Submittable: Submittable<B, SecondaryLevel, OutsideRenderPass>,

Execute commands from secondary buffers.

impl<'a, B, C, L> Encoder<'a, B, C, L> where
B: Backend,

pub fn level(&self) -> L where
L: Level,

Get encoder level.

pub fn copy_buffer<impl IntoIterator>(
&mut self,
src: &<B as Backend>::Buffer,
dst: &<B as Backend>::Buffer,
regions: impl IntoIterator
) where
C: Supports<Transfer>,
impl IntoIterator: IntoIterator<Item = BufferCopy>,

Copy buffer regions. src and dst can be the same buffer or alias in memory. But regions must not overlap. Otherwise resulting values are undefined.

pub fn copy_buffer_to_image<impl IntoIterator>(
&mut self,
src: &<B as Backend>::Buffer,
dst: &<B as Backend>::Image,
dst_layout: Layout,
regions: impl IntoIterator
) where
C: Supports<Transfer>,
impl IntoIterator: IntoIterator<Item = BufferImageCopy>,

Copy buffer region to image subresource range.

pub fn copy_image<impl IntoIterator>(
&mut self,
src: &<B as Backend>::Image,
src_layout: Layout,
dst: &<B as Backend>::Image,
dst_layout: Layout,
regions: impl IntoIterator
) where
C: Supports<Transfer>,
impl IntoIterator: IntoIterator<Item = ImageCopy>,

Copy image regions.

pub fn dispatch(&mut self, x: u32, y: u32, z: u32) where
C: Supports<Compute>,

Dispatch compute.

pub fn dispatch_indirect(
&mut self,
buffer: &<B as Backend>::Buffer,
offset: u64
) where
C: Supports<Compute>,

Dispatch indirect. Similar to dispatch except takes vertices and indices from buffer at specified offset. buffer must contain DispatchCommand at offset.