Module rendy::descriptor

Structs

DescriptorAllocator
Descriptor allocator. Can be used to allocate descriptor sets for any layout.

DescriptorRangeDesc
Set of descriptors of a specific type.

DescriptorRanges
Number of descriptors per type.

DescriptorRangesIter
Iterator over descriptor ranges.

DescriptorSet
Descriptor set from allocator.

DescriptorSetLayoutBinding
Information about the contents of and in which stages descriptors may be bound to a descriptor set at a certain binding point. Multiple DescriptorSetLayoutBindings are assembled into a DescriptorSetLayout, which is then allocated into a DescriptorSet using a DescriptorPool.

Enums

DescriptorType
DOC TODO: Grasping and remembering the differences between these types is a tough task. We might be able to come up with better names? Or even use tuples to describe functionality instead of coming up with fancy names.