Trait rendy::graph::Node

pub trait Node<B, T>: 'static + NodeSubmittable<'a, B> + Send + Sync + Debug where
    B: Backend,
    T: ?Sized,
    <Self::Desc as NodeDesc<B, T>>::Node == Self, {
    type Capability: Capability;
    type Desc: NodeDesc<B, T>;
    fn run(
        &'a mut self, 
        factory: &mut Factory<B>, 
        aux: &mut T, 
        frames: &'a Frames<B>
    ) -> Self::Submittables;

    unsafe fn dispose(self, factory: &mut Factory<B>, aux: &mut T);

    fn desc() -> Self::Desc
    where
        Self::Desc: Default,
    { ... }

    fn builder() -> NodeBuilder<B, T>
    where
        Self::Desc: Default,
    { ... }
}

The node is building block of the framegraph. Node defines set of resources and operations to perform over them. Read-only data for operations comes from auxiliary data source T.

Parameters

B - backend type. T - auxiliary data type.

Associated Types

type Capability: Capability

Capability required by node. Graph will execute this node on command queue that supports this capability level.

type Desc: NodeDesc<B, T>

Description type to instantiate the node.

Required Methods

fn run(
&'a mut self,
factory: &mut Factory<B>,
aux: &mut T,
frames: &'a Frames<B>
) -> Self::Submittables

Record commands required by node. Returned submits are guaranteed to be submitted within specified frame.

unsafe fn dispose(self, factory: &mut Factory<B>, aux: &mut T)

Dispose of the node.

Safety

Must be called after waiting for device idle.

Provided Methods

fn desc() -> Self::Desc where
Self::Desc: Default,

Desc creation. Convenient method if builder implements Default.

fn builder() -> NodeBuilder<B, T> where
Self::Desc: Default,

Builder creation.